Move the ProjectManager from widgets to views
